+++ src/test/kotlin/org/jetbrains/intellij/platform/gradle/tasks/BuildPluginTaskTest.kt
@@ -17,7 +17,7 @@ class BuildPluginTaskTest : IntelliJPluginTestBase() {
     private val sinceBuild: String
         get() {
             val version = Version.parse(intellijPlatformBuildNumber)
-            return "${version.major}.${version.minor}"
+            return "${version.major}"
         }
 
     @BeforeTest