+++ src/test/kotlin/org/jetbrains/intellij/platform/gradle/tasks/PatchPluginXmlTaskTest.kt
@@ -16,7 +16,7 @@ class PatchPluginXmlTaskTest : IntelliJPluginTestBase() {
     private val sinceBuild: String
         get() {
             val version = Version.parse(intellijPlatformBuildNumber)
-            return "${version.major}.${version.minor}"
+            return "${version.major}"
         }
 
     @Test